refactor and fix image bug
