+++ .gitignore
@@ -49,4 +49,4 @@ build-iPhoneSimulator/
 # unless supporting rvm < 1.11.0 or doing something fancy, ignore this:
 .rvmrc
 .DS_Store
-result.json
+artworks.json
